check if we're root before anything else to prevent weird error messages

+++ Makefile
@@ -83,6 +83,9 @@ all: initrd
 compile: check ccache setup $(packages) misc
 
 clean:
+	@if [ "`id -u`" != 0 ]; then \
+	echo "error: you cannot perform this operation unless you are root."; exit 1; \
+	fi
 	rm -rf $(BDIR) $(MDIR) initrd-$(CARCH).img.gz
 	rm -rf $(packages) vmlinuz-$(KERNELV)-fw$(KERNELREL)-$(CARCH) $(CDIR)/*
 	$(MAKE) -C src clean
